backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.1
google-generativeai==0.8.3
pytest==8.3.3
httpx==0.27.2
